refactor: centralize ID generation
Closes #195
